lekce_9/agent/pyproject.toml
[project]
name = "wazuh-agent"
version = "0.1.0"
description = "LangChain agent for Wazuh incident analysis and reporting"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-mcp-adapters>=0.1.12",
    "python-dotenv>=1.0.0",
    "reportlab>=4.0.0",
    "matplotlib>=3.8.0",
    "pandas>=2.2.0",
    "seaborn>=0.13.0",
    "Pillow>=10.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["main.py", "analyzer.py", "pdf_generator.py"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
